Extract shared getActiveExamVisibilityCondition helper
Resolves DEBT-366. Pure deduplication: the identical 14-line
private activeExamVisibilityCondition() methods on
DrizzleAttemptRepository and DrizzleQuestionRepository are
consolidated into a single getActiveExamVisibilityCondition
function in src/adapters/repositories/shared/active-exam-visibility.ts.

Both repos import the shared helper. No callsite shape changes.
No behavior change. Existing test suite is the safety net —
zero tests modified.
